+++ password/ChangeLog.md
@@ -4,7 +4,7 @@
 
 -   Split the main datatypes module (`Data.Password`) into a separate package: `password-types`.
     The new package just contains `Password`, `PasswordHash` and their helper functions/instances.
--   Adjusted entire `password` package to use the `Data.Password` from this new `password-types`.
+-   Adjusted entire `password` package to use the `Data.Password.Types` from this new `password-types`.
 
 ## 2.1.0.0